+++ src/main/scala/org/economicsl/auctions/multiunit/orderbooks/SortedAskOrders.scala
@@ -17,8 +17,8 @@ package org.economicsl.auctions.multiunit.orderbooks
 
 import java.util.UUID
 
+import org.economicsl.auctions.multiunit.orders.AskOrder
 import org.economicsl.auctions.{Quantity, Tradable}
-import org.economicsl.auctions.multiunit.AskOrder
 
 import scala.collection.immutable.TreeSet
