+++ src/test/scala/org/economicsl/auctions/multiunit/orderbooks/MatchedOrdersSpec.scala
@@ -3,6 +3,8 @@ package org.economicsl.auctions.multiunit.orderbooks
 import java.util.UUID
 
 import org.economicsl.auctions.multiunit
+import org.economicsl.auctions.multiunit.orders
+import org.economicsl.auctions.multiunit.orders.{LimitAskOrder, LimitBidOrder}
 import org.economicsl.auctions.{GoogleStock, Price, Quantity}
 import org.scalatest.{FlatSpec, Matchers}
 
@@ -14,12 +16,12 @@ class MatchedOrdersSpec extends FlatSpec with Matchers {
 
     // Create some multi-unit limit ask orders
     val issuer1 = UUID.randomUUID()
-    val order1 = multiunit.LimitAskOrder(issuer1, Price(10), Quantity(10), google)
+    val order1 = LimitAskOrder(issuer1, Price(10), Quantity(10), google)
 
     val issuer2 = UUID.randomUUID()
-    val order2 = multiunit.LimitBidOrder(issuer2, Price(15), Quantity(10), google)
+    val order2 = LimitBidOrder(issuer2, Price(15), Quantity(10), google)
 
-    val empty = MatchedOrders.empty[GoogleStock](multiunit.LimitAskOrder.ordering.reverse, multiunit.LimitBidOrder.ordering.reverse)
+    val empty = MatchedOrders.empty[GoogleStock](orders.LimitAskOrder.ordering.reverse, orders.LimitBidOrder.ordering.reverse)
     val (matched, optionalUnMatched) = empty + (issuer1 -> order1, issuer2 -> order2)
 
     assert(matched.nonEmpty)
@@ -31,13 +33,13 @@ class MatchedOrdersSpec extends FlatSpec with Matchers {
 
     // Create some multi-unit limit ask orders
     val issuer1 = UUID.randomUUID()
-    val order1 = multiunit.LimitAskOrder(issuer1, Price(10), Quantity(100), google)
+    val order1 = orders.LimitAskOrder(issuer1, Price(10), Quantity(100), google)
 
     val issuer2 = UUID.randomUUID()
-    val order2 = multiunit.LimitBidOrder(issuer2, Price(15), Quantity(10), google)
+    val order2 = orders.LimitBidOrder(issuer2, Price(15), Quantity(10), google)
 
     // Create an empty set of matched orders and add the newly created orders
-    val empty = MatchedOrders.empty[GoogleStock](multiunit.LimitAskOrder.ordering.reverse, multiunit.LimitBidOrder.ordering.reverse)
+    val empty = MatchedOrders.empty[GoogleStock](orders.LimitAskOrder.ordering.reverse, orders.LimitBidOrder.ordering.reverse)
     val (matched, optionalUnMatched) = empty + (issuer1 -> order1, issuer2 -> order2)
 
     assert(matched.nonEmpty)
@@ -53,13 +55,13 @@ class MatchedOrdersSpec extends FlatSpec with Matchers {
 
     // Create some multi-unit limit ask orders
     val issuer1 = UUID.randomUUID()
-    val order1 = multiunit.LimitAskOrder(issuer1, Price(10), Quantity(10), google)
+    val order1 = orders.LimitAskOrder(issuer1, Price(10), Quantity(10), google)
 
     val issuer2 = UUID.randomUUID()
-    val order2 = multiunit.LimitBidOrder(issuer2, Price(15), Quantity(100), google)
+    val order2 = orders.LimitBidOrder(issuer2, Price(15), Quantity(100), google)
 
     // Create an empty set of matched orders and add the newly created orders
-    val empty = MatchedOrders.empty[GoogleStock](multiunit.LimitAskOrder.ordering.reverse, multiunit.LimitBidOrder.ordering.reverse)
+    val empty = MatchedOrders.empty[GoogleStock](orders.LimitAskOrder.ordering.reverse, orders.LimitBidOrder.ordering.reverse)
     val (matched, optionalUnMatched) = empty + (issuer1 -> order1, issuer2 -> order2)
 
     assert(matched.nonEmpty)
