+++ src/test/scala/org/economicsl/auctions/multiunit/orderbooks/SortedAskOrdersSpec.scala
@@ -2,6 +2,8 @@ package org.economicsl.auctions.multiunit.orderbooks
 
 import java.util.UUID
 
+import org.economicsl.auctions.multiunit.orders
+import org.economicsl.auctions.multiunit.orders.LimitAskOrder
 import org.economicsl.auctions.{GoogleStock, Price, Quantity, multiunit}
 import org.scalatest.{FlatSpec, Matchers}
 
@@ -13,7 +15,7 @@ class SortedAskOrdersSpec extends FlatSpec with Matchers {
     // Create a multi-unit limit ask order
     val issuer = UUID.randomUUID()
     val google = new GoogleStock
-    val order = multiunit.LimitAskOrder(issuer, Price(10), Quantity(100), google)
+    val order = LimitAskOrder(issuer, Price(10), Quantity(100), google)
 
     // Create an empty order book and add the order
     val empty = SortedAskOrders.empty[GoogleStock]
@@ -37,13 +39,13 @@ class SortedAskOrdersSpec extends FlatSpec with Matchers {
 
     // Create some multi-unit limit ask orders
     val issuer1 = UUID.randomUUID()
-    val order1 = multiunit.LimitAskOrder(issuer1, Price(10), Quantity(10), google)
+    val order1 = orders.LimitAskOrder(issuer1, Price(10), Quantity(10), google)
 
     val issuer2 = UUID.randomUUID()
-    val order2 = multiunit.LimitAskOrder(issuer2, Price(5), Quantity(15), google)
+    val order2 = orders.LimitAskOrder(issuer2, Price(5), Quantity(15), google)
 
     val issuer3 = UUID.randomUUID()
-    val order3 = multiunit.LimitAskOrder(issuer3, Price(15), Quantity(100), google)
+    val order3 = orders.LimitAskOrder(issuer3, Price(15), Quantity(100), google)
 
     // Create an empty order book and add the orders
     val empty = SortedAskOrders.empty[GoogleStock]
